requirements.txt
# Vibecoder AgentOS Requirements
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Optional dependencies voor development
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0

# AI Services
anthropic==0.8.1
openai==1.93.0
python-dotenv==1.0.0

# Video Processing
yt-dlp==2023.12.30

# Intelligent Processing (NEW: Content-aware cropping)
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy>=1.21.0

# Queue System & Background Processing  
redis==5.0.1
rq==1.15.1
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# Database
sqlalchemy==2.0.23
sqlite3
psycopg2-binary==2.9.9
requests==2.31.0

# Utility dependencies
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4